geonli/requirements.txt
torch
torchvision
torchaudio
transformers
accelerate
pillow
fastapi
uvicorn
streamlit
requests
ultralytics
opencv-python